Updating the website
Haven't touched it in a while so things were out of date, added more download links to files, optimised the code a bit and redid the structure of the site to be better for users.

Still doesn't work on mobile so that's my next goal to fix, I also need to comment the CSS better and could possibally do some more styling on sub pages to make them look nicer.